Refactor entire JS API to work with bigints

+++ src/instructions/div.rs
@@ -28,7 +28,7 @@ impl Axecutor {
     fn instr_div_rm8(&mut self, i: Instruction) -> Result<(), AxError> {
         debug_assert_eq!(i.code(), Div_rm8);
 
-        let ax = self.reg_read_16(AX);
+        let ax = self.reg_read_16(AX) as u16;
 
         let op = self.instruction_operand(i, 0)?;
         let src_val = match op {
@@ -46,8 +46,8 @@ impl Axecutor {
 
         let (quotient, remainder) = (ax / src_val, ax % src_val);
 
-        self.reg_write_8(AL, quotient as u8);
-        self.reg_write_8(AH, remainder as u8);
+        self.reg_write_8(AL, quotient as u8 as u64);
+        self.reg_write_8(AH, remainder as u8 as u64);
 
         Ok(())
     }
@@ -77,8 +77,8 @@ impl Axecutor {
 
         let (quotient, remainder) = (dst_val / src_val, dst_val % src_val);
 
-        self.reg_write_16(AX, quotient as u16);
-        self.reg_write_16(DX, remainder as u16);
+        self.reg_write_16(AX, quotient as u16 as u64);
+        self.reg_write_16(DX, remainder as u16 as u64);
 
         Ok(())
     }
@@ -108,8 +108,8 @@ impl Axecutor {
 
         let (quotient, remainder) = (dst_val / src_val, dst_val % src_val);
 
-        self.reg_write_32(EAX, quotient as u32);
-        self.reg_write_32(EDX, remainder as u32);
+        self.reg_write_32(EAX, quotient as u32 as u64);
+        self.reg_write_32(EDX, remainder as u32 as u64);
 
         Ok(())
     }
