+++ src/instructions/imul.rs
@@ -55,7 +55,7 @@ impl Axecutor {
 
         let (result, overflow) = src_value.overflowing_mul(imm_value);
 
-        self.reg_write_16(dest_op.into(), result as u16);
+        self.reg_write_16(dest_op.into(), result as u16 as u64);
 
         if overflow {
             self.set_flags_u8(FLAG_CF | FLAG_OF, 0, 0);
@@ -90,7 +90,7 @@ impl Axecutor {
 
         let (result, overflow) = src_value.overflowing_mul(imm_value);
 
-        self.reg_write_32(dest_op.into(), result as u32);
+        self.reg_write_32(dest_op.into(), result as u32 as u64);
 
         if overflow {
             self.set_flags_u8(FLAG_CF | FLAG_OF, 0, 0);
@@ -160,7 +160,7 @@ impl Axecutor {
 
         let (result, overflow) = src_value.overflowing_mul(imm_value);
 
-        self.reg_write_16(dest_op.into(), result as u16);
+        self.reg_write_16(dest_op.into(), result as u16 as u64);
 
         if overflow {
             self.set_flags_u8(FLAG_CF | FLAG_OF, 0, 0);
@@ -195,7 +195,7 @@ impl Axecutor {
 
         let (result, overflow) = src_value.overflowing_mul(imm_value);
 
-        self.reg_write_32(dest_op.into(), result as u32);
+        self.reg_write_32(dest_op.into(), result as u32 as u64);
 
         if overflow {
             self.set_flags_u8(FLAG_CF | FLAG_OF, 0, 0);
@@ -259,7 +259,7 @@ impl Axecutor {
 
         let result = (dst_val as i16).wrapping_mul(src_val as i16);
 
-        self.reg_write_16(AX, result as u16);
+        self.reg_write_16(AX, result as u16 as u64);
 
         self.set_flags_u8(
             if result >> 7 != 0 && result >> 7 != -1 {
@@ -296,8 +296,8 @@ impl Axecutor {
 
         let result = (dst_val as i32).wrapping_mul(src_val as i32);
 
-        self.reg_write_16(AX, result as u16);
-        self.reg_write_16(DX, (result >> 16) as u16);
+        self.reg_write_16(AX, result as u16 as u64);
+        self.reg_write_16(DX, (result >> 16) as u16 as u64);
 
         self.set_flags_u16(
             if result >> 15 != 0 && result >> 15 != -1 {
@@ -334,8 +334,8 @@ impl Axecutor {
 
         let result = (dst_val as i64).wrapping_mul(src_val as i64);
 
-        self.reg_write_32(EAX, result as u32);
-        self.reg_write_32(EDX, (result >> 32) as u32);
+        self.reg_write_32(EAX, result as u32 as u64);
+        self.reg_write_32(EDX, (result >> 32) as u32 as u64);
 
         self.set_flags_u32(
             if result >> 31 != 0 && result >> 31 != -1 {
@@ -413,7 +413,7 @@ impl Axecutor {
 
         let result = (dst_val as i32).wrapping_mul(src_val as i32);
 
-        self.reg_write_16(dest_op.into(), result as u16);
+        self.reg_write_16(dest_op.into(), result as u16 as u64);
 
         self.set_flags_u16(
             if result >> 15 != 0 && result >> 15 != -1 {
@@ -453,7 +453,7 @@ impl Axecutor {
 
         let result = (dst_val as i64).wrapping_mul(src_val as i64);
 
-        self.reg_write_32(dest_op.into(), result as u32);
+        self.reg_write_32(dest_op.into(), result as u32 as u64);
 
         self.set_flags_u32(
             if result >> 31 != 0 && result >> 31 != -1 {
