+++ src/instructions/test.rs
@@ -42,13 +42,13 @@ impl Axecutor {
         let src_val = match src {
             Operand::Register(r) => self.reg_read_8(r),
             _ => fatal_error!("Invalid source operand {:?} for TEST r/m8, r8", src),
-        };
+        } as u8;
 
         let dest_val = match dest {
             Operand::Register(r) => self.reg_read_8(r),
             Operand::Memory(m) => self.mem_read_8(self.mem_addr(m))?,
             _ => fatal_error!("Invalid destination operand {:?} for TEST r/m8, r8", dest),
-        };
+        } as u8;
 
         let result = dest_val & src_val;
 
@@ -68,13 +68,13 @@ impl Axecutor {
         let src_val = match src {
             Operand::Register(r) => self.reg_read_16(r),
             _ => fatal_error!("Invalid source operand {:?} for TEST r/m16, r16", src),
-        };
+        } as u16;
 
         let dest_val = match dest {
             Operand::Register(r) => self.reg_read_16(r),
             Operand::Memory(m) => self.mem_read_16(self.mem_addr(m))?,
             _ => fatal_error!("Invalid destination operand {:?} for TEST r/m16, r16", dest),
-        };
+        } as u16;
 
         let result = dest_val & src_val;
 
@@ -94,13 +94,13 @@ impl Axecutor {
         let src_val = match src {
             Operand::Register(r) => self.reg_read_32(r),
             _ => fatal_error!("Invalid source operand {:?} for TEST r/m32, r32", src),
-        };
+        } as u32;
 
         let dest_val = match dest {
             Operand::Register(r) => self.reg_read_32(r),
             Operand::Memory(m) => self.mem_read_32(self.mem_addr(m))?,
             _ => fatal_error!("Invalid destination operand {:?} for TEST r/m32, r32", dest),
-        };
+        } as u32;
 
         let result = dest_val & src_val;
 
@@ -190,7 +190,7 @@ impl Axecutor {
             Operand::Register(r) => self.reg_read_8(r),
             Operand::Memory(m) => self.mem_read_8(self.mem_addr(m))?,
             _ => fatal_error!("Invalid destination operand {:?} for TEST r/m8, imm8", dest),
-        };
+        } as u8;
 
         let result = dest_val & src_val;
 
@@ -231,7 +231,7 @@ impl Axecutor {
                 "Invalid destination operand {:?} for TEST r/m16, imm16",
                 dest
             ),
-        };
+        } as u16;
 
         let result = dest_val & src_val;
 
@@ -262,7 +262,7 @@ impl Axecutor {
                 "Invalid destination operand {:?} for TEST r/m32, imm32",
                 dest
             ),
-        };
+        } as u32;
 
         let result = dest_val & src_val;
